Rename slices folder to modules
