+++ modules/xkcd/xkcd.py
